refactor(codegen): merge the leaf chain predicates and layer schema.ts
Both remaining review findings.

The five one-export modules holding the small builder-chain and handler predicates become one
`functions/chain.ts`. One file per exported function was the wrong granularity at the leaf: they
are one vocabulary — a feeder wanting `procedureHandler` almost always wants `isDatabaseAccessor`
in the same breath — and five 10-line modules made the call sites harder to read than the code
they imported. `functions/` goes from 17 files to 13. The coarse splits are unchanged.

`schema.ts` (1,526 lines) becomes four modules along the dependency layering its own graph
already had, verified acyclic with no back-references before the cut:

    schema/internal/properties.ts     144   object-literal readers, a leaf with no outward refs
    schema/internal/table-builder.ts  835   the table builder parser
    schema/internal/extensions.ts     507   schema-extension merging
    schema/index.ts                    76   discoverSchema

`schema/index.ts` keeps `./discover/schema` resolving, so no importer changed and the public
surface is untouched.

Also drops the symbol resolution `resolvesToImportedName` was doing per callee. It runs on every
`.use(...)` argument of every chain, and reading an import alias needs no type checker — scanning
the file's import declarations answers it syntactically, which is also what keeps the detector
working under degraded type info. A per-file cache was tried and removed: ts-morph reuses the
`SourceFile` object when a file is overwritten, so the cache served the previous content's
aliases — a wrong answer about whether a procedure declares a policy, for a saving this scan does
not need.
