--- packages/codegen/src/discover/functions/inline-handler.ts
@@ -1,12 +1,0 @@
-import type { ArrowFunction, FunctionExpression } from "ts-morph";
-import { Node } from "ts-morph";
-
-/** A function whose body we can inspect — an inline arrow or function expression handler. */
-type InspectableHandler = ArrowFunction | FunctionExpression;
-
-/** The inline arrow/function-expression handler at `argument`, or `undefined` when it isn't one. */
-const inlineHandler = (argument: Node | undefined): InspectableHandler | undefined =>
-    argument !== undefined && (Node.isArrowFunction(argument) || Node.isFunctionExpression(argument)) ? argument : undefined;
-
-export { inlineHandler };
-export type { InspectableHandler };
