--- packages/codegen/src/discover/functions/is-database-accessor.ts
@@ -1,7 +1,0 @@
-import { Node } from "ts-morph";
-
-/** True when `receiver` is the database accessor: `ctx.db` (property named `db`) or a bare `db`. */
-const isDatabaseAccessor = (receiver: Node): boolean =>
-    (Node.isPropertyAccessExpression(receiver) && receiver.getName() === "db") || (Node.isIdentifier(receiver) && receiver.getText() === "db");
-
-export default isDatabaseAccessor;
