--- packages/codegen/src/discover/functions/procedure-handler.ts
@@ -1,32 +1,0 @@
-import type { CallExpression } from "ts-morph";
-import { Node } from "ts-morph";
-
-import type { InspectableHandler } from "./inline-handler";
-import { inlineHandler } from "./inline-handler";
-
-/**
- * The inline handler function of a classified procedure call, or `undefined` when
- * it isn't inspectable. The terminal call's first argument is either the handler
- * function directly (`query(async ({ ctx }) => …)` / `c.use(…).query(handler)`) or
- * an object literal carrying it under a `handler` property (`query({ args, handler })`)
- * — both surface forms are handled. The companion to `classifyProcedureCall`:
- * classify the call, then pull out the body to inspect.
- */
-const procedureHandler = (initializer: CallExpression): InspectableHandler | undefined => {
-    const argument = initializer.getArguments()[0];
-    const direct = inlineHandler(argument);
-
-    if (direct !== undefined) {
-        return direct;
-    }
-
-    if (argument === undefined || !Node.isObjectLiteralExpression(argument)) {
-        return undefined;
-    }
-
-    const property = argument.getProperty("handler");
-
-    return property !== undefined && Node.isPropertyAssignment(property) ? inlineHandler(property.getInitializer()) : undefined;
-};
-
-export default procedureHandler;
