+++ packages/codegen/src/discover/raw-row-returns.ts
@@ -4,12 +4,9 @@ import { Node, SyntaxKind } from "ts-morph";
 import { singleHopInitializer } from "../argument-taint";
 import type { RawRowReturnIR } from "../ir";
 import { listLunoraSourceFiles, lunoraRelativePath } from "./ast";
-import chainHasStep from "./functions/chain-has-step";
-import chainUsesWrappedCall from "./functions/chain-uses-wrapped-call";
+import type { InspectableHandler } from "./functions/chain";
+import { chainHasStep, chainUsesWrappedCall, isDatabaseAccessor, procedureHandler } from "./functions/chain";
 import { classifyProcedureCall } from "./functions/classify-procedure-call";
-import type { InspectableHandler } from "./functions/inline-handler";
-import isDatabaseAccessor from "./functions/is-database-accessor";
-import procedureHandler from "./functions/procedure-handler";
 
 /** `ctx.db` read methods that hand back a whole row (or array of rows): the by-id `get` and the `findFirst`/`findMany` family. */
 const ROW_READ_METHODS = new Set(["findFirst", "findFirstOrThrow", "findMany", "get"]);
